# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Job Application Automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all scraped jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, applied, failed, skipped, filtered_non_us]
        - name: platform
          in: query
          schema:
            type: string
            enum: [linkedin, indeed, greenhouse, lever, handshake]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJobStatus
      tags: [jobs]
      summary: Update job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStatusUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Remove job from queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/preview:
    get:
      operationId: getJobPreview
      tags: [jobs]
      summary: Get job with generated cover letter preview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job with cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPreview"

  # Applications
...
  /greenhouse/companies:
    get:
      operationId: getGreenhouseCompanies
      tags: [scheduler]
      summary: Get the Greenhouse company seed list
      responses:
        "200":
          description: Company list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GreenhouseCompanyListResponse"
    put:
      operationId: updateGreenhouseCompanies
      tags: [scheduler]
      summary: Replace the full Greenhouse company list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GreenhouseCompanyListResponse"
      responses:
        "200":
          description: Updated company list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GreenhouseCompanyListResponse"
    post:
      operationId: addGreenhouseCompany
      tags: [scheduler]
      summary: Add a company to the Greenhouse seed list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GreenhouseCompany"
      responses:
        "201":
          description: Company added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GreenhouseCompany"
  /greenhouse/companies/{slug}:
    delete:
      operationId: removeGreenhouseCompany
      tags: [scheduler]
      summary: Remove a company from the Greenhouse seed list
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]

  # Proxies
  /proxies:
    get:
      operationId: listProxies
      tags: [proxies]
      summary: List all configured proxies
      responses:
        "200":
          description: Proxy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proxy"
    post:
      operationId: createProxy
      tags: [proxies]
      summary: Add a new proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyInput"
      responses:
        "201":
          description: Created proxy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
  /proxies/{id}:
    patch:
      operationId: updateProxy
      tags: [proxies]
      summary: Update a proxy entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyInput"
      responses:
        "200":
          description: Updated proxy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
        "404":
          description: Not found
    delete:
      operationId: deleteProxy
      tags: [proxies]
      summary: Remove a proxy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /proxies/{id}/test:
    post:
      operationId: testProxy
      tags: [proxies]
      summary: Test a proxy connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyTestResult"
  /proxies/rotation:
    get:
      operationId: getProxyRotationSettings
      tags: [proxies]
      summary: Get proxy rotation settings
      responses:
        "200":
          description: Rotation settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyRotationSettings"
    put:
      operationId: updateProxyRotationSettings
      tags: [proxies]
      summary: Update proxy rotation settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyRotationSettingsInput"
      responses:
        "200":
          description: Updated rotation settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyRotationSettings"

  # AI
  /ai/generate-cover-letter:
    post:
      operationId: generateCoverLetter
      tags: [ai]
      summary: Generate a cover letter for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterRequest"
      responses:
        "200":
          description: Generated cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterResponse"
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [linkedin, indeed, greenhouse, lever, handshake]
        jobTitle:
          type: string
        company:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        jobDescription:
          type: ["string", "null"]
        applyUrl:
          type: string
        scrapedAt:
          type: string
        status:
          type: string
          enum: [queued, applied, failed, skipped, filtered_non_us]
        coverLetterText:
          type: ["string", "null"]
        countryResolved:
          type: ["string", "null"]
      required: [id, platform, jobTitle, company, remote, applyUrl, scrapedAt, status]
    JobPreview:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        coverLetter:
          type: ["string", "null"]
      required: [job]
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
      required: [jobs, total]
    JobStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [queued, applied, failed, skipped, filtered_non_us]
      required: [status]
...
    GreenhouseCompany:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
      required: [slug, name]
    GreenhouseCompanyListResponse:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: "#/components/schemas/GreenhouseCompany"
      required: [companies]
...
    Proxy:
      type: object
      properties:
        id:
          type: integer
        label:
          type: ["string", "null"]
        host:
          type: string
        port:
          type: string
        username:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        protocol:
          type: string
          enum: [http, https, socks5]
        enabled:
          type: boolean
        lastTestedAt:
          type: ["string", "null"]
        lastTestStatus:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, host, port, protocol, enabled, createdAt]
    ProxyInput:
      type: object
      properties:
        label:
          type: string
        host:
          type: string
        port:
          type: string
        username:
          type: string
        password:
          type: string
        protocol:
          type: string
          enum: [http, https, socks5]
        enabled:
          type: boolean
      required: [host, port]
    ProxyTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]
        ip:
          type: ["string", "null"]
      required: [success, message]
    ProxyRotationSettings:
      type: object
      properties:
        enabled:
          type: boolean
        strategy:
          type: string
          enum: [round-robin, random, least-used]
        rotateEvery:
          type: integer
        activeProxyId:
          type: ["integer", "null"]
      required: [enabled, strategy, rotateEvery]
    ProxyRotationSettingsInput:
      type: object
      properties:
        enabled:
          type: boolean
        strategy:
          type: string
          enum: [round-robin, random, least-used]
        rotateEvery:
          type: integer
    CoverLetterRequest:
      type: object
      properties:
        jobId:
          type: integer
      required: [jobId]
    CoverLetterResponse:
      type: object
      properties:
        coverLetter:
          type: string
        cached:
          type: boolean
      required: [coverLetter, cached]